Fix circular import, todo: check functions in find/locate (swapped)
