+++ circletracking/find.py
@@ -3,6 +3,7 @@ from __future__ import (absolute_import, division, print_function,
 import six
 import numpy as np
 from numpy.testing import assert_allclose
+import skimage
 try:
     from skimage.filters import threshold_otsu
 except ImportError:
@@ -11,25 +12,32 @@ from skimage.measure import find_contours
 import pandas
 
 from .algebraic import fit_ellipse
-from .locate import locate_multiple_disks
-from .refine import refine_multiple
-
 
 def find_disks(image, size_range, number_of_disks=100):
     """
-    Find circular particles in the image
-    :param size_range:
+    Locate blobs in the image by using a Laplacian of Gaussian method
     :rtype : pandas.DataFrame
-    :param n:
     :return:
     """
-    blobs = locate_multiple_disks(image, size_range, number_of_disks)
+    number_of_disks = int(np.round(number_of_disks))
+    radii = np.linspace(size_range[0], size_range[1],
+                        num=min(abs(size_range[0] - size_range[1]) * 2.0, 30),
+                        dtype=np.float)
 
-    if blobs.empty:
-        return pandas.DataFrame(columns=['r', 'y', 'x', 'dev'])
+    # Find edges
+    edges = skimage.feature.canny(image)
+    circles = skimage.transform.hough_circle(edges, radii)
 
-    return refine_multiple(image, blobs, size_range,
-                           num_points_circle=number_of_disks)
+    fit = pandas.DataFrame(columns=['r', 'y', 'x', 'accum'])
+    for radius, hough_circle in zip(radii, circles):
+        peaks = skimage.feature.peak_local_max(hough_circle, threshold_rel=0.5,
+                                               num_peaks=number_of_disks)
+        accumulator = hough_circle[peaks[:, 0], peaks[:, 1]]
+        fit = pandas.concat([fit, pandas.DataFrame(data={'r': [radius] * peaks.shape[0], 'y': peaks[:, 0], 'x': peaks[:, 1], 'accum': accumulator}) ], ignore_index=True)
+
+        fit = merge_hough_same_values(fit, number_of_disks)
+
+        return fit
 
 def find_ellipse(image, mode='ellipse_aligned', min_length=24):
     """ Thresholds the image, finds the longest contour and fits an ellipse
@@ -115,3 +123,38 @@ def find_ellipsoid(image3d, center_atol=None):
                         err_msg='Found centers have inconsistent values.')
 
     return zr, yr, xr, zc, yc, xc
+
+
+def merge_hough_same_values(data, number_to_keep=100):
+    """
+
+    :param data:
+    :return:
+    """
+    while True:
+        # Rescale positions, so that pairs are identified below a distance
+        # of 1. Do so every iteration (room for improvement?)
+        positions = data[['x', 'y']].values
+        mass = data['accum'].values
+        duplicates = scipy.spatial.cKDTree(positions, 30).query_pairs(
+            np.mean(data['r']), p=2.0, eps=0.1)
+        if len(duplicates) == 0:
+            break
+        to_drop = []
+        for pair in duplicates:
+            # Drop the dimmer one.
+            if np.equal(*mass.take(pair, 0)):
+                # Rare corner case: a tie!
+                # Break ties by sorting by sum of coordinates, to avoid
+                # any randomness resulting from cKDTree returning a set.
+                dimmer = np.argsort(np.sum(positions.take(pair, 0), 1))[0]
+            else:
+                dimmer = np.argmin(mass.take(pair, 0))
+            to_drop.append(pair[dimmer])
+        data.drop(to_drop, inplace=True)
+
+    # Keep only brightest n circles
+    data = data.sort_values(by=['accum'], ascending=False)
+    data = data.head(number_to_keep)
+
+    return data
